[수정] 클린 아키텍쳐로 폴더구조 개선
