+++ UJeongApp/APP/Platform/Politician/UseCases/PoliticianUseCase.swift
@@ -1,0 +1,25 @@
+//
+//  PoliticianUseCase.swift
+//  UJeongApp
+//
+//  Created by Hamlit Jason on 2023/01/02.
+//
+
+import Foundation
+import RxSwift
+
+protocol PoliticianUseCase {
+    func fetchPolitician() -> Observable<Politician>
+}
+
+final class PoliticianUseCaseImpl: PoliticianUseCase {
+    private let repository: PoliticianRepsitory
+    
+    init(repository: PoliticianRepsitory) {
+        self.repository = repository
+    }
+    
+    func fetchPolitician() -> Observable<Politician> {
+        return repository.getPoliticians()
+    }
+}
