+++ UJeongApp/Core/Models/Policician.swift
@@ -1,0 +1,14 @@
+//
+//  Policician.swift
+//  UJeongApp
+//
+//  Created by Hamlit Jason on 2023/01/02.
+//
+
+import Foundation
+
+struct Politician {
+    let name: String // 정치인 이름
+    let party: String // 정당
+    let constituency: String // 지역구
+}
